♻️ (Refactoring) - Reorganize some files + Do some refactoring
