+++ src/app/Sidebar/utils/border/removeBorder.ts
