+++ src/app/Sidebar/utils/border/updateBorder.ts
